Using fakesu and apt-get now
Instead of packaging all the sources for dbus and dbus-glib in the
buildpack for download, I am using a buildpack to install a chroot jail
on Heroku, and then installing dbus and dbus-glib directly from the apt
repository.

Once the dependencies are installed, python-dbus is downloaded from its
website and it will attempt to be installed.

I suspect the configure script will break, for unknown reasons. The
fakeroot jail makes it such that lots of dependencies become missing,
like gcc. I don't undersand it yet. I will push this to a test
environment on Heroku and read the error messages.